+++ src/components/Explorer/Explorer.stories.tsx
@@ -23,7 +23,11 @@ const Template: Story<{
     createPermission?: string,
     editPermission?: string,
     deletePermission?: string,
-    details?: object,
+    details?: {
+        page: string,
+        params: unknown
+    },
+    buttons?: [],
     children?: React.ReactNode,
     layout?: string
 }> = ({
@@ -33,30 +37,36 @@ const Template: Story<{
     details,
     children,
     layout = 'basic',
+    buttons,
     ...props
 }) => {
     const toast = React.useRef(null);
-    const show = action => data => toast.current.show({
+    const show = props => params => toast.current.show({
         severity: 'success',
         summary: 'Submit',
-        detail: <pre>{JSON.stringify({action, data}, null, 2)}</pre>
+        detail: <pre>{JSON.stringify({...props, params}, null, 2)}</pre>
     });
     return (
         <>
-            <Toast ref={toast} />
             <div style={{height: 'fit-content', display: 'flex', flexDirection: 'column'}}>
                 <Explorer
                     fetch={fetchItems}
                     keyField='id'
                     resultSet='items'
                     name='items'
+                    methods={{
+                        async 'portal.customization.get'() {
+                            return {};
+                        },
+                        'explorer.submit': show({method: 'explorer.submit'})
+                    }}
                     schema={{
                         properties: {
                             id: {
-                                action: show('action')
+                                action: show({action: 'action'})
                             },
                             name: {
-                                action: show('action'),
+                                action: show({action: 'action'}),
                                 title: 'Name',
                                 filter: true,
                                 sort: true
@@ -87,7 +97,7 @@ const Template: Story<{
                         }
                     }}
                     subscribe={updateItems}
-                    onCustomization={show('customization')}
+                    onCustomization={show({handler: 'onCustomization'})}
                     details={details}
                     filter={{}}
                     cards={{
@@ -121,7 +131,7 @@ const Template: Story<{
                             }]
                         },
                         toolbar: {
-                            widgets: [{
+                            widgets: buttons || [{
                                 title: 'Create',
                                 permission: createPermission,
                                 action: () => {}
@@ -129,12 +139,12 @@ const Template: Story<{
                                 title: 'Edit',
                                 permission: editPermission,
                                 enabled: 'current',
-                                action: show('edit')
+                                action: show({action: 'edit'})
                             }, {
                                 title: 'Delete',
                                 permission: deletePermission,
                                 enabled: 'selected',
-                                action: show('delete')
+                                action: show({action: 'delete'})
                             }]
                         }
                     }}
@@ -162,6 +172,7 @@ const Template: Story<{
                     {children}
                 </Explorer>
             </div>
+            <Toast ref={toast} />
         </>
     );
 };
@@ -184,8 +195,20 @@ Children.args = {
 export const Details = Template.bind({});
 Details.args = {
     ...Children.args,
+    middleware: [
+        _store => next => action => (action.type === 'portal.component.get')
+            ? Promise.resolve(function Details({value: {preview: {current, pagination}}}) {
+                return <div>
+                    <div><Text>Name</Text>: {current.name}</div>
+                    <div><Text>Size</Text>: {current.size}</div>
+                    <div><Text>Records</Text>: {pagination.recordsTotal}</div>
+                </div>;
+            })
+            : next(action)
+    ],
+    table: {selectionMode: 'single'},
     details: {
-        name: 'Name'
+        page: 'details'
     }
 };
 
@@ -202,6 +225,36 @@ ActionPermissions.args = {
     createPermission: 'forbidden',
     editPermission: 'granted',
     deletePermission: 'forbidden'
+};
+
+/* eslint-disable no-template-curly-in-string */
+export const Submit = Template.bind({});
+Submit.args = {
+    ...Basic.args,
+    buttons: [{
+        title: 'Submit id',
+        enabled: 'single',
+        method: 'explorer.submit',
+        params: '${id}'
+    }, {
+        title: 'Submit current',
+        enabled: 'current',
+        method: 'explorer.submit',
+        params: '${current}'
+    }, {
+        title: 'Submit selected',
+        enabled: 'selected',
+        method: 'explorer.submit',
+        params: '${selected}'
+    }, {
+        title: 'Submit template',
+        method: 'explorer.submit',
+        enabled: 'current',
+        params: {
+            id: '${id}',
+            size: '${current.size}'
+        }
+    }]
 };
 
 export const DateTimeFilter = Template.bind({});