+++ src/components/Explorer/Explorer.types.ts
@@ -52,7 +52,10 @@ export interface Props {
     /**
      * Fields to show in the details pane.
      */
-    details?: object;
+    details?: {
+        page: string,
+        params?: unknown
+    };
     toolbar?: false | WidgetReference[];
     filter?: Record<string, unknown>;
     params?: Record<string, unknown>;
@@ -72,6 +75,8 @@ export interface Props {
     customization?: Customization,
     paramsCard?: string,
     onFieldChange?: string,
+    value?: unknown,
+    onChange?: (params: object) => void,
     onCustomization?: (customization: {component: {componentId: string, componentConfig: Customization}}) => Promise<object>,
 }