+++ src/components/Form/index.tsx
@@ -153,7 +153,7 @@ const Form: ComponentProps = ({
         {devTool ? <DevTool control={control} placement="top-right" /> : null}
         {toast}
         {toolbarElement}
-        <form {...rest} onSubmit={submit} className={clsx('grid col align-self-start', classes.form, className)}>
+        <div {...rest} className={clsx('grid col align-self-start', classes.form, className)}>
             {!!errorList.length && <div className='col-12'>{errorList}</div>}
             {layoutState.visibleCards.map((id1, level1) => {
                 const nested = [].concat(id1);
@@ -219,7 +219,7 @@ const Form: ComponentProps = ({
                     drop
                 />
             </div>}
-        </form>
+        </div>
     </>);
 };