+++ src/components/ThumbIndex/ThumbIndex.types.ts
@@ -11,6 +11,9 @@ export interface Props extends React.HTMLAttributes<HTMLDivElement> {
     name?: string;
     items: Item[];
     orientation?: 'left' | 'top',
+    type?: 'steps' | 'tabs' | 'thumbs',
+    loading?: boolean,
+    trigger?: React.MouseEventHandler,
     onFilter: (filter: unknown) => void
 }