+++ src/components/prime/index.tsx
@@ -1,4 +1,4 @@
-import { Button as PrimeButton, type ButtonProps } from 'primereact/button';
+import { Button as PrimeButton, type ButtonProps as PrimeButtonProps } from 'primereact/button';
 import { Calendar } from 'primereact/calendar';
 import { Card as PrimeCard, type CardProps } from 'primereact/card';
 import { AutoComplete as PrimeAutoComplete, type AutoCompleteProps } from 'primereact/autocomplete';
@@ -8,6 +8,7 @@ import type { FileUploadProps } from 'primereact/fileupload';
 import { DataTable as PrimeDataTable } from 'primereact/datatable';
 import React from 'react';
 import Text from '../Text';
+import Permission from '../Permission';
 
 export { Calendar } from 'primereact/calendar';
 export { CascadeSelect } from 'primereact/cascadeselect';
@@ -36,6 +37,7 @@ export { Ripple } from 'primereact/ripple';
 export { SelectButton } from 'primereact/selectbutton';
 export { Skeleton } from 'primereact/skeleton';
 export { Splitter, SplitterPanel } from 'primereact/splitter';
+export { Steps } from 'primereact/steps';
 export { TabMenu } from 'primereact/tabmenu';
 export { TabPanel, TabView } from 'primereact/tabview';
 export { Toast } from 'primereact/toast';
@@ -75,15 +77,19 @@ export const DateRange = props => {
 
 export const Card = ({title, ...props}: CardProps) =>
     <PrimeCard title={title && <Text>{title}</Text>} {...props}/>;
-export const Button = ({children, ...props}: ButtonProps) =>
-    <PrimeButton {...props}>{children && <span className='p-button-label p-c'><Text>{children}</Text></span>}</PrimeButton>;
+
+export type ButtonProps = PrimeButtonProps & Partial<Pick<Parameters<typeof Permission>[0], 'permission'>>
+export const Button = ({children, permission, ...props}: ButtonProps) => {
+    const button = <PrimeButton {...props}>{children && <span className='p-button-label p-c'><Text>{children}</Text></span>}</PrimeButton>;
+    return (permission == null) ? button : <Permission permission={permission}>{button}</Permission>;
+};
 export const DataTable = ({emptyMessage = 'No results found', ...props}: DataTableProps) =>
     <PrimeDataTable emptyMessage={emptyMessage && <Text>{emptyMessage}</Text>} {...props}/>;
 export const AutoComplete = React.forwardRef<PrimeAutoComplete, AutoCompleteProps & {methods: unknown, autocomplete?: string}>(
     function AutoComplete({methods, autocomplete, ...props}, ref) {
         const [suggestions, setSuggestions] = React.useState();
         const complete = React.useCallback(
-            async event => methods && autocomplete && setSuggestions((await methods[autocomplete](event)).suggestions), [methods, autocomplete]);
+            async({query}) => methods && autocomplete && setSuggestions((await methods[autocomplete]({query})).suggestions), [methods, autocomplete]);
         return <PrimeAutoComplete {...props} completeMethod={complete} suggestions={suggestions} ref={ref}/>;
     }
 );